Move loader into .template
